+++ src/Hacks/UnlockButtons.cpp
@@ -1,18 +1,51 @@
-/*#ifndef GEODE_IS_MACOS
+#ifndef GEODE_IS_MACOS
 
 #include <Geode/Geode.hpp>
 #include <Geode/modify/CCSpriteGrayscale.hpp>
+#include <Geode/modify/CCSpriteWithHue.hpp>
 #include "../Client/Client.h"
 
 class $modify(CCSpriteGrayscale)
 {
     static CCSpriteGrayscale* createWithSpriteFrame(CCSpriteFrame* spr)
     {
-        if (Client::GetModuleEnabled("unlock-buttons"))
-            return as<CCSpriteGrayscale*>(CCSprite::createWithSpriteFrame(spr));
+        auto res = CCSpriteGrayscale::createWithSpriteFrame(spr);
 
-        return CCSpriteGrayscale::createWithSpriteFrame(spr);
+        auto spr2 = CCSprite::createWithSpriteFrame(spr);
+        spr2->setPosition(res->getContentSize() / 2);
+
+        res->addChild(spr2);
+        res->setOpacity(0);
+        
+        return res;
+    }
+
+    static void onModify(auto& self) {
+        std::vector<geode::Hook*> hooks;
+
+        auto it = self.m_hooks.begin();
+        std::advance(it, 0);
+
+        hooks.push_back(it->second.get());
+
+        Loader::get()->queueInMainThread([hooks] 
+        {
+            auto modu = Client::GetModule("unlock-buttons");
+
+            for (auto hook : hooks)
+            {
+                if (hook)
+                {
+                    hook->setAutoEnable(false);
+
+                    if (!modu->enabled)
+                        hook->disable();
+
+                    modu->hooks.push_back(hook);
+                }
+            }
+        });
     }
 };
 
-#endif*/
+#endif