+++ changelog.md
@@ -2,13 +2,16 @@
 
 - Added Conditional Auto-Retry
 - Added Auto Practice Mode
+- Added Unlock Buttons
 - Redesigned the button tab in the config menu, this also resulted in Hide Button being moved to the menu tab
 - Show Layout shows Hidden objects
+- Added an option to customize show layout bg and ground colours
 - Fixed Show Layout making the ground's white
 - Show Layout now disables pulse triggers
 - Fixed Show Triggers making the triggers disappear once they've been on screen
 - Fixed a bug that makes cataclysm and a few other levels impossible
 - Fixed Incompatibility with better pause and being able to click through some popups
+- Fixed Gradient Appearing white
 
 # 1.3.9